+++ src/cljs/cxengage_javascript_sdk/modules/sqs.cljs
@@ -45,14 +45,26 @@
           current-session-id (state/get-session-id)
           msg-type (or (:notification-type parsed-body)
                        (:type parsed-body))]
-      (if (not= (state/get-session-id) session-id)
-        (do #_(log :warn (str "Received a message from a different session than the current one."
-                                         "Current session ID: " current-session-id
-                                         " - Session ID on message received: " session-id
-                                         " Message type: " msg-type))
-            nil)
-        (do (delete-message-fn receipt-handle)
-            body)))))
+      ;; Three cases on handling SQS messages for sessions:
+      ;; First, no user session has started where the message returns
+      ;; to the queue until this SDK instance has a session.
+      ;; Second, messages from a previous session where the message
+      ;; is deleted and not processed by the SDK instance.
+      ;; Third, messages from this session where the message
+      ;; is deleted and processed by the SDK instance.
+      ;; Finally, since this session is no longer valid,
+      ;; the implicit case is  a newer session's messages
+      ;; returning to the queue; consequently,
+      ;; the newer session can delete dead messages in this queue as per
+      ;; the second use case above.
+      (when (nil? current-session-id)
+        nil)
+      (when (iu/uuid-came-before? session-id current-session-id)
+        (delete-message-fn receipt-handle)
+        nil)
+      (when (= (state/get-session-id) session-id)
+        (delete-message-fn receipt-handle)
+        body))))
 
 (defn delete-message*
   [sqs {:keys [queue-url]} receipt-handle]